+++ Operations/mafOpVOIDensity.cpp
@@ -185,7 +185,7 @@ void mafOpVOIDensity::OnEvent(mafEventBase *maf_event)
 			case ID_CHOOSE_SURFACE:
 			{
 				mafString title = _L("VOI surface");
-        mafEvent event(this,VME_CHOOSE,&title,(long)&mafOpVOIDensity::OutputSurfaceAccept);
+        mafEvent event(this,VME_CHOOSE,&title,(intptr_t)&mafOpVOIDensity::OutputSurfaceAccept);
 				mafEventMacro(event);
 				m_Surface = event.GetVme();
 				if(m_Surface == NULL)
